feat: implement support for more well-known-types
- google.protobuf.Struct (mapped to a JSON STRING)
- google.type.TimeOfDay (mapped to TIME)
- google.type.Date (mapped to DATE)
- google.type.LatLng (mapped to GEOGRAPHY)

The reason for using GEOGRAPHY as default for LatLng is because of the
clustering opportunities for tables that store locations as GEOGRAPHY.

Future SchemaOptions may override to store LatLng as a record instead.

Integration tested against public BigQuery datasets. The intention is to
identify datasets that exercise the features of this library, then
perform a load/marshal/unmarshal sequence and validate that the original
result is preserved.