feat: add support for lossy and non-lossy DATETIME conversion
google.protobuf.DateTime and DATETIME differ in that the protobuf
version can include a time zone or a time offset.

To ensure complete lossless saving and loading, a custom RECORD type is
used by default, that saves all the time offset fields.

In the case where the time offset is not used or not needed, an optional
schema flag `UseDateTimeWithoutOffset` can be toggled, leading to
DATETIME being used as the column type for google.protobuf.DateTime.