feat: add schema option UseOneofFields
When set, if the message contains any oneof fields, a BigQuery STRING field will be created
for each oneof group. It will then contain the name of the oneof that is set. This gives
support for having oneof groups where one of the fields in the group is an empty message.

This commit also fixes the behavior of InferSchema to skip fields that are empty messages
since that is not supported by BigQuery.